Add SQLite database support for tracking posted articles
Magticom publishes some news with a date in a long future (like a month from now), and they stay on top the whole time. This prevented other news from being detected, as they appeared below the "future" one.

This change will save all the news the script seen in a sqlite db, and parse all the news on the page to find the new ones.